services/pipeline-runner/requirements.txt
mysql-connector-python
minio
pyarrow
pandas
ddtrace
oracledb
pymssql
